+++ apis/r/NEWS.md
@@ -1,3 +1,7 @@
+# 1.10.2
+
+* Port resume-ingest mode from Python to R
+
 # 1.10.1
 
 * This release contains a single Python-only bug fix
@@ -11,6 +15,7 @@
 * Add support for reading `*m` and `*p` layers from `SOMAExperimentAxisQuery`
 * Add support for blockwise iteration
 * Make `reopen()` a public method for all `TileDBObjects`
+* Add support for resume-mode in `write_soma()`
 
 # 1.7.0